pyproject.toml
[project]
name = "flask-celery-poc"
version = "0.1.0"
description = "Flask + Celery + Redis production-ready app"
authors = [
    {name = "ArkiLovesProgramming",email = "betaweiwang@gmail.com"}
]
license = "MIT"
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "alembic (==1.12.0)",
    "amqp (==5.3.1)",
    "billiard (==4.2.1)",
    "bcrypt (==4.0.1)",
    "blinker (==1.9.0)",
    "cachelib (==0.9.0)",
    "celery (==5.3.1)",
    "cffi (==1.17.1)",
    "click (==8.2.1)",
    "click-didyoumean (==0.3.1)",
    "click-plugins (==1.1.1.2)",
    "click-repl (==0.3.0)",
    "colorama (==0.4.6)",
    "connexion[flask,uvicorn] (==3.0.2)",
    "flask (==2.3.3)",
    "flask-caching (==2.0.2)",
    "flask-sqlalchemy (==3.0.5)",
    "gevent (==25.5.1)",
    "greenlet (==3.2.4)",
    "itsdangerous (==2.2.0)",
    "jinja2 (==3.1.6)",
    "jsonschema (==4.20.0)",
    "kombu (==5.3.2)",
    "mako (==1.3.10)",
    "markupsafe (==3.0.2)",
    "openapi-spec-validator (==0.7.1)",
    "packaging (==25.0)",
    "passlib[bcrypt] (==1.7.4)",
    "prompt-toolkit (==3.0.51)",
    "pycparser (==2.22)",
    "PyJWT (==2.8.0)",
    "python-dateutil (==2.9.0.post0)",
    "python-dotenv (>=1.0.0)",
    "python-json-logger (==2.0.7)",
    "redis (==5.0.0)",
    "setuptools (==80.9.0)",
    "six (==1.17.0)",
    "sqlalchemy (==2.0.44)",
    "swagger-ui-bundle (==1.1.0)",
    "typing-extensions (==4.14.1)",
    "tzdata (==2025.2)",
    "vine (==5.1.0)",
    "wcwidth (==0.2.13)",
    "werkzeug (==3.1.3)",
    "zope-event (==5.1.1)",
    "zope-interface (==7.2)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[dependency-groups]
dev = [
    "pytest (>=9.0.1,<10.0.0)",
    "pytest-cov (>=7.0.0,<8.0.0)",
    "flake8 (>=7.0.0,<8.0.0)"
]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra --maxfail=1 --cov=app --cov-report=term-missing"
testpaths = [
    "tests"
]
filterwarnings = [
    "ignore::DeprecationWarning"
]
